+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update `amazon-chime-sdk-js` dependency to v3 beta.
 - Update the `compilerOptions.target` in `tsconfig.json` from `es5` to `ES2015 (ES6)`.
 - Rename the `global` property of `DefaultTheme` Interface to `globalStyle` to avoid conflict with reserved keyword `global`.
+- Change `Versioning.ts` to read from Git instead of manually hard-coded. This is the same behavior in 
+  `amazon-chime-sdk-js`.
 
 ### Removed